// src/app/shared/interfaces.ts
export interface User {
  email: string;
  password: string;
  returnSecureToken?: boolean;
}

export interface FbAuthResponse {
  idToken: string;
  expiresIn: string;
}

export interface Post {
  id?: string;
  title: string;
  content: string;
  author: string;
  date: Date;
}

export class FbCreateResponse {
  name: string;
}